new encoding settings per resolution
bit_rate check for videoCanBeCopied.
If the bitrate of an video exceeds the threshold for this particular
resolution, the video is deemed not copyable and will be encoded.

crf rate can now be set on a per resolution basis.
the preset setting can now be set on a per resolution basis.